hw-04

+++ model/index.js
@@ -1,34 +1,40 @@
 const { Post } = require('../db/postModel')
 
-const listContacts = async () => {
-  const contacts = await Post.find({})
+const listContacts = async (userId) => {
+  const contacts = await Post.find({ owner: userId })
   return contacts
 }
-const getContactById = async (contactId) => {
-  const contact = await Post.findById(contactId)
+const getContactById = async (contactId, userId) => {
+  const contact = await Post.find({ _id: contactId, owner: userId })
+  if (Object.keys(contact).length === 0) { return undefined }
   return contact
 }
 
-const removeContact = async (contactId) => {
-  const contact = await getContactById(contactId)
+const removeContact = async (contactId, userId) => {
+  const contact = await getContactById(contactId, userId)
+  if (!contact) { return undefined }
   await Post.findByIdAndRemove(contactId)
   return contact
 }
 
-const addContact = async (body) => {
+const addContact = async (body, userId) => {
   const { name, email, phone } = body
-  const post = new Post({ name, email, phone })
+  const post = new Post({ name, email, phone, owner: userId })
   await post.save()
   return body
 }
 
-const updateContact = async (contactId, body) => {
+const updateContact = async (contactId, body, userId) => {
+  const contact = await getContactById(contactId, userId)
+  if (!contact) { return undefined }
   const { name, email, phone } = body
   await Post.findByIdAndUpdate(contactId, { $set: { name, email, phone } })
   return body
 }
 
-const updateFavoriteContact = async (contactId, body) => {
+const updateFavoriteContact = async (contactId, body, userId) => {
+  const contact = await getContactById(contactId, userId)
+  if (!contact) { return undefined }
   const { name, email, phone, favorite } = body
   let isFavorite = true
   if (favorite === void 0 || favorite === true) { isFavorite = true } else { isFavorite = false }
